+++ product/src/main/java/store/_0982/product/application/dto/GroupPurchaseUpdateCommand.java
@@ -1,7 +1,5 @@
 package store._0982.product.application.dto;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
@@ -10,7 +8,7 @@ public record GroupPurchaseUpdateCommand(
         int maxQuantity,
         String title,
         String description,
-        int discountedPrice,
+        Long discountedPrice,
         OffsetDateTime startDate,
         OffsetDateTime endDate,
         UUID productId
